todos/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = []
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.10"
aerich = "0.5.3"
aiofiles = "0.6.0"
bcrypt = "3.2.0"
click = "7.1.2"
fastapi = "0.63.0"
itsdangerous = "1.1.0"
pillow = "8.2.0"
pyjwt = "1.7.1"
python-dotenv = "0.17.0"
python-multipart = "0.0.5"
tortoise-orm = "0.17.2"
twilio = "6.57.0"

[tool.poetry.group.dev.dependencies]
pyproject-flake8 = "6.0.0"
mypy = "0.812"
asynctest = "0.13.0"
httpx = "0.18.0"
pytest = "6.2.3"
pytest-asyncio = "0.15.1"

[tool.flake8]
exclude = ["./migrations", "venv", ".git", "__pycache__", ".mypy_cache", ".pytest_cache"]
ignore = ["E402", "W503"]
max-line-length = 79

[tool.mypy]
ignore_missing_imports = true

[tool.aerich]
tortoise_orm = "app.settings._POSTGRES_TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
